+++ CHANGES.md
@@ -1,12 +1,11 @@
 # videojs-record changelog
 
-## 4.3.0 - unreleased
+## 4.3.0 - 2021/03/08
 
 - Fix controlbar flickering in Mobile Safari (#413)
 - Fix hiding picture-in-picture button (#486)
 - Fix stretched output for image-only mode when aspect ratio
   of player is not the same as camera media constraints (#511)
-- Fix `saveAs` file-extension for extended mimetypes (#553)
 - Bump required version for videojs-wavesurfer (3.6.0 or newer)